+++ CHANGES.md
@@ -81,6 +81,9 @@ To be released.
     `[string | URL | URLPattern, Temporal.Duration | Temporal.DurationLike][]`
     (was `[string | URL | URLPattern, Temporal.Duration][]`).
 
+ - Adds the `-A`/`--authorized-fetch` flag to the `fedify inbox` command. [[#229], [#472] By Lee ByeongJun]
+
+[#229]: https://github.com/fedify-dev/fedify/issues/229
 [#280]: https://github.com/fedify-dev/fedify/issues/280
 [#366]: https://github.com/fedify-dev/fedify/issues/366
 [#376]: https://github.com/fedify-dev/fedify/issues/376